pyproject.toml
[project]
name = "rcl"
version = "0.1.0"
description = "Reflective Context Learning — optimization primitives for context-space agent adaptation"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0",
    "anthropic>=0.40",
    "openai>=1.60",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "numpy>=1.24",
]

[project.optional-dependencies]
rewardbench2 = ["datasets>=2.0"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["rcl*", "benchmarks*"]
